+++ server/collusic-be/src/main/java/com/collusic/collusicbe/service/RequestProjectService.java
@@ -1,7 +1,6 @@
 package com.collusic.collusicbe.service;
 
 import com.collusic.collusicbe.domain.requestproject.RequestProjectRepository;
-import com.collusic.collusicbe.web.dto.RequestProjectDto;
 import com.collusic.collusicbe.web.dto.RequestProjectSaveRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,8 +18,7 @@ public class RequestProjectService {
     @Transactional
     public Long save(RequestProjectSaveRequestDto requestProjectSaveRequestDto) throws IOException {
         String uploadFilePath = s3Service.upload(requestProjectSaveRequestDto.getMultipartFile(), "static");
-        RequestProjectDto requestProjectDto = requestProjectSaveRequestDto.getRequestProjectDto();
-        requestProjectDto.setUploadFilePath(uploadFilePath);
-        return requestProjectRepository.save(requestProjectDto.toEntity()).getId();
+        requestProjectSaveRequestDto.setUploadFilePath(uploadFilePath);
+        return requestProjectRepository.save(requestProjectSaveRequestDto.toEntity()).getId();
     }
 }
