+++ server/collusic-be/src/main/resources/application.properties
@@ -12,8 +12,8 @@ spring.jpa.properties.hibernate.jdbc.batch_size=10
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 
-spring.servlet.multipart.max-file-size=10MB
-spring.servlet.multipart.max-request-size=10MB
+spring.servlet.multipart.max-file-size=15MB
+spring.servlet.multipart.max-request-size=15MB
 
 spring.profiles.include=aws
